Refactor into reuseable component
